Feat: 글 상세 페이지 구현 및 피드백 반영하여 수정
