+++ kudog/lib/pages/home/ViewHotPage.dart
